refacor: prevalidate() method.
